requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
bootstrap4==0.1.0
dj-database-url==0.5.0
Django==3.0.4
django-bootstrap4==1.1.1
django-crispy-forms==1.8.1
django-debug-toolbar==2.2
django-filter==2.2.0
django-heroku==0.3.1
django-tables2==2.2.1
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.0.0
psycopg2==2.8.4
python-dotenv==0.12.0
pytz==2019.3
soupsieve==1.9.5
sqlparse==0.3.0
whitenoise==5.0.1
